TST: Add VibrationsData dict tests w/ diff indices
The to/fromdict tests have been parametrized to cover the test case
outlined in #1650 and grouped into a test class to avoid parametrizing
unrelated tests and for organization.

Note that the n2_vibdata fixture now strips rows from the Hessian
according to those indices that are not specified by the `indices`
fixture. In the case that indices=None (the default for the indices
parameter in VibrationsData.__init__), no modifications are made. This
is performed within the fixture since the fixture is implicitly
parametrized with the indices fixture, and the dimensions of the
Hessians must be (indices, 3, indices, 3). However, this means that
when testing with constraints, the Hessian must be modified
manually.** This is the case with the current tests covering
constraints-test_fixed_atoms, test_constrained_atoms.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>